Interactivity Fix
prohibit the computation for equivalent classes when one of the
equivalenceset has been processed.

Fixes the bug that there are 4 Arrows between 2 Classes if one of them
has 4 equivalence classes.

Fix 2:
Compute the dependencytree from scratch.
Deleting two nodes that have the same single superclass and the same
single subclass results in a single arrow from super to sub instead of 2
arrows.