Complete Restructure
Code is now structured as close as possible to the design document.
Problems:
- Layouting has gone all weird. The size of the nodes is being set but
the width is too small.
- Cannot get changes to be reflected (add super-/subclasses etc) without
a complete layout pass (the updateView method now only adds or removes
nodes instead of completely rebuilding)
- Name bugs are still present
- Multiple main windows should work but has not been tested.

Notes:
- Colors are now set without a new layout pass
- The App class tracks main window instances and the associated
viewManager and drawingService objects.
- All yFiles code has been moved into teo.isgci.yfiles (still need to
double check that it isn't being used elsewhere).

Questions:
- What is teo.isg?
- What does the 'createMaps' parameter do and do we need to implement
it?